+++ public/js/src/module/translator.js
@@ -44,7 +44,7 @@ initialize = new Promise( function( resolve, reject ) {
             fallbackLng: 'en',
             joinArrays: '\n',
             backend: {
-                loadPath: settings.basePath + '/locales/__lng__/translation-combined.json',
+                loadPath: settings.basePath + '/locales/build/__lng__/translation-combined.json',
             },
             load: 'languageOnly',
             lowerCaseLng: true,